% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCVgene.R
\name{nCVgene}
\alias{nCVgene}
\title{Evaluate robustness of clock genes in population scale data}
\usage{
nCVgene(inputD, cgenes)
}
\arguments{
\item{inputD}{a data frame. The first column is the gene symbol, and other columns are samples.
One row per gene. Do not accept data frame with duplicate gene symbols.}

\item{cgenes}{a character vector. The targeted clock genes for returning the nCV values.}
}
\value{
A data frame with calculated nCV values of targeted clock genes.
}
\description{
This function calculate nCV values of targeted clock genes. The nCV values can indicate
the robustness of clock genes. Higher nCV value means more robust of oscillation.
Before applying this function, please do the clock functionality test using \code{nCVnet}.
If the \code{nCVnet} output significant p-value, it is good to run \code{nCVgene}.
}
\examples{
## calculate the nCV values of targeted clock genes
cgenes = c("ARNTL", "CLOCK", "NPAS2", "CRY1", "NR1D1", "CIART", "DBP", "PER1", "CRY2", "PER2")
ncvD = nCVgene(inputD = nCVegD, cgenes)
ncvD
}
