% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nCVnet.R
\name{nCVnet}
\alias{nCVnet}
\title{Evaluate functionality of circadian clock network in population scale data}
\usage{
nCVnet(inputD, benchD, hs = TRUE, nperm = 1000, seedN = 10)
}
\arguments{
\item{inputD}{a data frame. The first column is the gene symbol, and other columns are samples.
One row per gene. Do not accept data frame with duplicate gene symbols.}

\item{benchD}{a data frame. The expression correlation values of paired clock genes. The}

\item{hs}{a logical value. If the samples are from human, set \code{"hs"} as \code{TRUE}.
If the samples are from mouse, set \code{hs} as \code{FALSE}.}

\item{nperm}{a numeric value. The number of permutations used to calculate the p-value. Default is 1000.}

\item{seedN}{a numeric value. Specify the seeds for random sampling.}
}
\value{
A list containing the following components:
\tabular{rl}{
zstat    \tab  the permutation test results of \code{nCVnet}
npermV   \tab  a vector of Z-statistic (Mantel test) of permutations
cmatrix  \tab  a matrix of correlation values of paired clock genes using the input data
}
}
\description{
This function can test whether there is a functional clock network in population scale data.
If the p-value of this test is significant, the next step is to get the nCV values of target genes.
}
\examples{
## test the whether there is a functional clock network running in the example data
testD = nCVnet(inputD = nCVegD, benchD = mClockD)
testD$zstat
}
